Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int x;
-   scanf("%d", &x);
-   printf("You entered: %d",x);
+    float x;
+   scanf("%f", &x);
+   printf("You entered: %f",x);
    return 0;
 }
